fix: :bug: 显示优化&持续性测试；修复「AIDummy生成函数在响应游戏刻时不正常」的问题

+++ src/batr/game/main/GlobalGameVariables.ts
@@ -11,6 +11,6 @@ export const TICK_TIME_MS: number = TICK_TIME_S * 1000;
 // Timer:A delay lower than 20 milliseconds instanceof not recommended
 
 /** 用于「武器无冷却」模式中的「最小冷却」 */
-export const TOOL_MIN_CD: uint = TPS / 8;
+export const TOOL_MIN_CD: uint = uint(TPS / 8);
 /** 抛射体在玩家面前生成的距离 */
 export const PROJECTILES_SPAWN_DISTANCE: number = 0.55;
